v1

+++ app/Http/Controllers/Admin/RolesController.php
@@ -11,6 +11,7 @@ use App\Models\Role;
 use Gate;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
+Use Alert;
 
 class RolesController extends Controller
 {
@@ -37,6 +38,7 @@ class RolesController extends Controller
         $role = Role::create($request->all());
         $role->permissions()->sync($request->input('permissions', []));
 
+        Alert::success(trans('global.flash.success'), trans('global.flash.created'));
         return redirect()->route('admin.roles.index');
     }
 
@@ -44,9 +46,15 @@ class RolesController extends Controller
     {
         abort_if(Gate::denies('role_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+        if($role->id == 1){
+            Alert::error('cant edit');
+            return back();
+        }
+
         $permissions = Permission::pluck('title', 'id');
 
         $role->load('permissions');
+
 
         return view('admin.roles.edit', compact('permissions', 'role'));
     }
@@ -56,6 +64,7 @@ class RolesController extends Controller
         $role->update($request->all());
         $role->permissions()->sync($request->input('permissions', []));
 
+        Alert::success(trans('global.flash.success'), trans('global.flash.updated'));
         return redirect()->route('admin.roles.index');
     }
 
@@ -72,8 +81,13 @@ class RolesController extends Controller
     {
         abort_if(Gate::denies('role_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+        if($role->id == 1){
+            Alert::error('cant delete');
+            return back();
+        }
         $role->delete();
 
+        Alert::success(trans('global.flash.success'), trans('global.flash.deleted'));
         return back();
     }
